Implement SysDir read/write; use in delete+rename
The new function make_sysdir_array() creates an array of
afs_dv_t entries from the SysDir contents.

In this array the type is set to 0 for a delete.
The rename writes the new filename into the entry found.

Then the SysDir contents is regenerated from the array
and written back to the in-memory disk image.

Writing the file tries to allocate new pages, if the size of the
file does not fit in the pages already allocated to it.