Implement truncate; remove afs_fileinfo_t::pages
The fileinfo pages were not used, so just get rid of them.

Now truncate() does something and you can overwrite existing files.

TODO: Verify that all statistics in the afs_fileinfo_t of a file
are updated. Also the leader page is not yet updated for operations
which change the file.