Create a class AltoFS to contain everything
Instead of a set of C style exported functions, now altofs.cpp/h
define the class AltoFS which contains everything.

I.e. the code is now (a bit) C++-ified.

Using std::vector<afs_dv> for the list of directory entries to get rid
of memmove at least for this part of the code.

Add create_file(const char* path) and make it create a new SysDir
entry, as well as a child node in the root_dir->child array.

Creating files works in principle, but writing to currently
them seems to lose the first page. There's a problem with the leader
page or with what page reference is inserted in the SysDir.

TODO: simplify SysDir handling; perhaps have a cache of SysDir entries,
i.e. a std::map<std::string,afs_dv> and always use this when modifying
SysDir in some way. Right now the SysDir file is read/written all the
time.