This is like writing sort of a fsck.alto
The discrepancies between the file labels and their info,
compared to the bit table bits, is one thing. Finally the
SysDir has to be consulted to distinguish allocated from
unallocated files and not e.g. mark bits as 'used' when
the file itself was delete from the SysDir.

This turns out to be a little difficult and before it is
solved, writing to the disk will almost always really damage it.