Fix a stupid bug in endian_t; fsck.alto improved
This was a really stupid bug in endian_t. Of course in a union of a
16bit and two 8bit values, the latter have to be an array in order
to occupy the lower and more significant bytes. DUH!

As a consequence all code relying on m_little was wrong.

Now this is fixed and reading the files from
a) the disk image by identifying valid leader pages
b) the SysDir file and its valid (and deleted) entries
works much better. No more actually deleted files in the
mount point.