No swap on getword/putword
Judging from the amount of files which _seem_ to need fixing
when opening a disk image, it looks like swapping the word
in getword() and putword() is wrong. Or it is wrong for little
endian :)

The fuse-alto.cpp functions now pass the path to the altofs.cpp
code rather than using find_fileinfo() on their side to see if
a file exists or not.